chore: move pages style to correct place
