+++ packages/marketplace/src/components/pages/app-detail/__styles__/app-permission-content.scss
