+++ packages/marketplace/src/components/pages/app-detail/client/app-install-confirmation.tsx
@@ -3,7 +3,7 @@ import { useHistory } from 'react-router'
 import { History } from 'history'
 import { useDispatch, useSelector } from 'react-redux'
 import { AppDetailModel } from '@reapit/foundations-ts-definitions'
-import appPermissionContentStyles from '@/styles/pages/app-permission-content.scss?mod'
+import appPermissionContentStyles from '../__styles__/app-permission-content.scss?mod'
 import { Button, ModalV2, GridFourCol, GridFourColItem, Content, ModalPropsV2 } from '@reapit/elements'
 import { installApp } from '@/actions/installations'
 import { fetchAppDetail } from '@/actions/apps'
