+++ packages/marketplace/src/components/pages/app-detail/client/app-uninstall-confirmation.tsx
@@ -4,7 +4,7 @@ import { History } from 'history'
 import { Dispatch } from 'redux'
 import { useDispatch, useSelector } from 'react-redux'
 import { AppDetailModel } from '@reapit/foundations-ts-definitions'
-import appPermissionContentStyles from '@/styles/pages/app-permission-content.scss?mod'
+import appPermissionContentStyles from '../__styles__/app-permission-content.scss?mod'
 import { fetchAppDetail } from '@/actions/apps'
 import { Button, ModalV2, ModalPropsV2 } from '@reapit/elements'
 import { uninstallApp } from '@/actions/installations'
