+++ packages/marketplace/src/components/pages/help/__styles__/help.scss
@@ -1,5 +1,5 @@
-@import '../base/colors.scss';
-@import '../base/browser.scss';
+@import '@/styles/base/colors.scss';
+@import '@/styles/base/browser.scss';
 
 .wrapBoxContent {
   height: 100%;
