+++ packages/marketplace/src/components/pages/help/help-item-list.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import styles from '@/styles/pages/help.scss?mod'
+import styles from './__styles__/help.scss?mod'
 import { Button, GridFourCol, GridThreeColItem, H5, Section, LevelRight, Content } from '@reapit/elements'
 
 export interface HelpItem {
