+++ packages/marketplace/src/components/pages/login/__styles__/login.scss
@@ -1,4 +1,4 @@
-@import '../base/colors.scss';
+@import '@/styles/base/colors.scss';
 
 .container {
   min-width: 100vw;
