+++ packages/marketplace/src/components/pages/login/login.tsx
@@ -2,7 +2,7 @@ import * as React from 'react'
 import { Button, Level, FlexContainerBasic, Section, notification } from '@reapit/elements'
 import { getDefaultRoute } from '@/utils/auth-route'
 import messages from '@/constants/messages'
-import loginStyles from '@/styles/pages/login.scss?mod'
+import loginStyles from './__styles__/login.scss?mod'
 import logoImage from '@/assets/images/reapit-graphic.jpg'
 import connectImage from '@/assets/images/reapit-connect.png'
 import { reapitConnectBrowserSession } from '@/core/connect-session'
