+++ packages/marketplace/src/components/pages/settings/__styles__/settings.scss
@@ -1,0 +1,5 @@
+@import '@/styles/base/colors.scss';
+
+.user-info-section {
+  color: $grey;
+}
