--- packages/marketplace/src/styles/pages/settings.scss
@@ -1,5 +1,0 @@
-@import '../../styles/base/colors.scss';
-
-.user-info-section {
-  color: $grey;
-}
