--- packages/marketplace/src/styles/pages/welcome.scss
@@ -1,20 +1,0 @@
-.graphic {
-  width: 100%;
-  height: auto;
-}
-
-.container {
-  display: flex;
-  margin: 0 auto;
-  width: 100%;
-
-  @media (min-width: 769px) {
-    width: 80%;
-  }
-}
-
-.content {
-  @media (min-width: 769px) {
-    max-width: 375px;
-  }
-}
